requirements.txt
# Core utilities
requests>=2.32.0
tqdm>=4.66.1
numpy>=1.25.0
pandas>=2.1.0
pdfplumber>=0.9.0
faiss-cpu>=1.7.4
pyarrow

# Transformers & embeddings
transformers>=4.42.0
sentence-transformers>=2.2.2

# Google Gemini / generative AI
google-generativeai>=0.3.1
google-genai
torch
transformers
sentence-transformers
faiss-cpu
pdfplumber
requests
pandas
numpy
tqdm
scikit-learn
more-itertools